DB-3793:
- Generate d.ts files on build with rollup plugin
- Remove rollup config, use build.rollupOptions in vite config
