+++ packages/nextjs-kit/index.ts
@@ -1,0 +1,2 @@
+export * from './src/components';
+export * from './src/lib';
