+++ packages/nextjs-kit/package.json
@@ -11,18 +11,11 @@
   },
   "author": "@pantheon-systems",
   "files": [
-    "dist",
-    "src/sdk.{js,ts}"
+    "dist"
   ],
-  "main": "./dist/umd/nextjs-kit.umd.js",
-  "module": "./dist/es/*-es.js",
-  "exports": {
-    ".": {
-      "import": "./dist/es",
-      "require": "./dist/umd/nextjs-kit.umd.js"
-    },
-    "./*": "./dist/es/*.js"
-  },
+  "main": "./dist/nextjs-kit.umd.js",
+  "module": "./dist/nextjs-kit.es.js",
+  "types": "./dist/index.d.ts",
   "prettier": {
     "singleQuote": true,
     "arrowParens": "avoid",
@@ -36,8 +29,8 @@
     "*.{md,mdx}": "prettier --write"
   },
   "scripts": {
-    "build": "tsc && rollup -c",
-    "watch": "tsc && rollup -c -w",
+    "build": "tsc && vite build",
+    "watch": "tsc && vite build --watch",
     "typedoc": "typedoc",
     "test": "vitest run",
     "test:watch": "vitest",
@@ -49,13 +42,17 @@
     "lint-staged": "lint-staged"
   },
   "devDependencies": {
+    "@tailwindcss/typography": "^0.5.2",
     "@testing-library/react": "12.1.5",
     "@vitejs/plugin-react": "^2.0.0",
+    "autoprefixer": "^10.4.4",
     "c8": "^7.12.0",
     "next": "12.1.6",
+    "postcss": "^8.4.12",
     "react": "17.0.2",
     "react-dom": "17.0.2",
     "tailwindcss": "3.1.6",
+    "vite-plugin-dts": "^1.4.1",
     "vitest": "^0.18.0"
   },
   "peerDependencies": {
