+++ packages/nextjs-kit/src/index.ts
@@ -1,7 +1,2 @@
-// entrypoint for the umd rollup build - note the default exports
-import Header from './components/header';
-import Footer from './components/footer';
-import sortChar from './lib/sortChar';
-import sortDate from './lib/sortDate';
-
-export default { Header, Footer, sortChar, sortDate };
+export * from './components';
+export * from './lib';
