+++ packages/nextjs-kit/tsconfig.json
@@ -1,6 +1,6 @@
 {
   "extends": "../../tsconfig.json",
   "compilerOptions": {
-    "jsx": "react-jsx"
-  },
+    "jsx": "react-jsx",
+  }
 }
