--- packages/nextjs-kit/umd-entrypoint.ts
@@ -1,3 +1,0 @@
-import * as NextjsKit from './src';
-
-export default NextjsKit;
